server/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
langchain==0.1.11
langchain-groq==0.1.1
langchain-community==0.0.24
pydantic==2.6.1
python-multipart==0.0.6
gunicorn==21.2.0